Add files via upload
A sales_data table is created to store sales transaction details, including customer name, product name, unit price, quantity, and total price. To automate the calculation of total_price, a PL/pgSQL function calc_tot_price is defined to compute it as the product of unit price (sales) and quantity. A trigger calc_tot_insert is set to execute this function before an insert operation on the table. This setup ensures accurate and automatic computation of the total price for each transaction, minimizing errors and manual effort.